+++ CHANGES.rst
@@ -188,6 +188,13 @@ Bug Fixes
   - Fixed the AttributeError exception which was raised when using
     :func:`astropy.wcs.WCS.footprint_to_file`. [#1912]
 
+  - Fixed a NameError exception which was raised when using
+    :func:`astropy.wcs.validate` or the `wcslint` script. [#2053]
+
+  - Fixed a bug where named WCSes may be erroneously reported as ``'
+    '`` when using :func:`astropy.wcs.validate` or the `wcslint`
+    script. [#2053]
+
 - Misc
 
   - There are a number of improvements to make `astropy` work better